refactor(server): extract macro calendar → macro.mjs
